Optimize concrete readWordFromBytes
